requirements.txt
selenium
pandas
openpyxl
matplotlib
request
sqlite3












